Created Goblin NPC Behavior
Goblin can wander to and from four points randomly.
At a certain distance the goblin will chase the player when he approaches and attack at a certain distance.
The player can kill the goblin with his sword with one hit. (We can change this to make it harder if we want)
Patrol, Chase, Attack, and Death Animations